% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrashCodes.R
\name{CrashCodes}
\alias{CrashCodes}
\title{CrashCodes: affiche un "crâne" ASCII si on entre "Rstudio"}
\usage{
CrashCodes(x)
}
\arguments{
\item{x}{caractère d'entrée (doit être un seul élément)}
}
\value{
NA_character_ ou, invisiblement, le vecteur de lignes du dessin ASCII
}
\description{
Si l'argument n'est pas exactement "Rstudio", renvoie NA_character_.
Si l'argument vaut "Rstudio", affiche une art ASCII et retourne invisiblement
le texte (vecteur de lignes).
}
\examples{
CrashCodes("Rstudio")
CrashCodes("autre")  # renvoie NA
}
