% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianeParCategorie.R
\name{MedianeParCategorie}
\alias{MedianeParCategorie}
\title{Médiane d'une variable par catégorie}
\usage{
MedianeParCategorie(var, par, dans)
}
\arguments{
\item{var}{Colonne numérique à résumer (non-quotée), ex. Note}

\item{par}{Colonne de groupement (non-quotée), ex. Couleur}

\item{dans}{data.frame contenant les colonnes (non-quoté), ex. Exercice_Serie_2}
}
\value{
data.frame avec une ligne par catégorie et la médiane correspondante
(colonne nommée \code{med_<var>}).
}
\description{
Calcule la médiane d'une variable numérique pour chaque modalité d'une
variable catégorielle, en ignorant systématiquement les valeurs manquantes.
Usage naturel : MedianeParCategorie(Note, par = Couleur, dans = Exercice_Serie_2)
}
\examples{
# Médiane des notes par couleur
# MedianeParCategorie(Note, par = Couleur, dans = Exercice_Serie_2)

}
