% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode statistique}
\usage{
Mode(x, na_rm = TRUE, ties = c("all", "first", "random"))
}
\arguments{
\item{x}{vecteur}

\item{na_rm}{TRUE pour ignorer les NA}

\item{ties}{"all","first","random"}
}
\value{
valeur(s) modale(s)
}
\description{
Mode statistique
}
