% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moyenne.R
\name{Moyenne}
\alias{Moyenne}
\title{Moyenne (équivalent de mean)}
\usage{
Moyenne(x, ...)
}
\arguments{
\item{x}{Vecteur numérique.}

\item{...}{Arguments additionnels passés à \code{\link[base:mean]{base::mean()}}, par ex. \code{na.rm = TRUE}.}
}
\value{
Un scalaire numérique (la moyenne).
}
\description{
Calcule la moyenne d'un vecteur numérique.
}
\examples{
Moyenne(c(1, 2, 3))
Moyenne(c(1, NA, 3), na.rm = TRUE)
}
