% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoyenneParCategorie.R
\name{MoyenneParCategorie}
\alias{MoyenneParCategorie}
\title{Moyenne d'une variable par catégorie}
\usage{
MoyenneParCategorie(var, par, dans)
}
\arguments{
\item{var}{colonne numérique à résumer (non-quotée), ex. Note}

\item{par}{colonne de groupement (non-quotée), ex. Couleur}

\item{dans}{data.frame (non-quoté), ex. Exercice_Serie_2}
}
\value{
data.frame (une ligne par catégorie, colonne moy_*)
}
\description{
Calcule la moyenne d'une variable numérique pour chaque modalité d'une
variable catégorielle, en ignorant les NA.
Usage : MoyenneParCategorie(Note, par = Couleur, dans = Exercice_Serie_2)
}
