% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode statistique}
\usage{
Mode(x, na_rm = TRUE, ties = c("all", "first", "random"))
}
\arguments{
\item{x}{vecteur.}

\item{na_rm}{TRUE pour ignorer les NA.}

\item{ties}{Stratégie en cas d'ex æquo : \code{"all"}, \code{"first"}, ou \code{"random"}.}
}
\value{
La/les valeur(s) modale(s) (même type que \code{x}).
}
\description{
Calcule la/les valeur(s) modale(s) d'un vecteur.
}
\examples{
Mode(c(1, 2, 2, 3))                # 2
Mode(c(1, 1, 2, 2), ties = "all")  # c(1, 2)
Mode(c("a", "b", "b", "a"), ties = "first")  # "a"
}
