% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moyenne.R
\name{Moyenne}
\alias{Moyenne}
\title{Moyenne (équivalent de mean)}
\usage{
Moyenne(x, ...)
}
\arguments{
\item{x}{Vecteur numérique.}

\item{...}{Passé à base::mean (ex. na.rm = TRUE).}
}
\value{
Un scalaire numérique (la moyenne).
}
\description{
Moyenne (équivalent de mean)
}
